Backend/app/schemas.py
from pydantic import BaseModel

class ProductBase(BaseModel):
    name:str
    description:str | None=None
    price: float
    stock_quantity : int

class ProductCreate(ProductBase):
    pass
class ProductUpdate(ProductBase):
    pass
class ProductResponse(ProductBase):
    product_id:int

    class Config:
        from_attributes = True


class UserCreate(BaseModel):
    username:str
    email:str
    password:str

class UserResponse(BaseModel):
    userid:int
    username:str
    email:str

    class Config:
        from_attributes = True

class Token(BaseModel):
    access_token :str
    token_type :str