Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pytest 
httpx
pytest-asyncio
python-dotenv
python-jose
passlib[bcrypt] 
python-multipart